Use document create hook to set locale
